pyproject.toml
[project]
name = "linkedin-agent"
version = "0.1.0"
description = ""
authors = [
    {name = "Andy Pang",email = "chunkhai96@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4"
dependencies = [
    "langchain (>=0.3.23,<0.4.0)",
    "langchain-community (>=0.3.21,<0.4.0)",
    "langchain-google-genai (>=2.1.2,<3.0.0)",
    "langgraph (>=0.3.27,<0.4.0)",
    "linkedin-api-client (>=0.3.0,<0.4.0)",
    "duckduckgo-search (>=8.0.0,<9.0.0)"
]

[tool.poetry]
packages = [{include = "linkedin_agent", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
